refactor: move transactions payload
